LU-8174 misc: name open file handles as such
In a number of places in the code, rename variables from "*_handle"
or "*_fh" to "*_open_handle" so that it is clear this is referencing
an open file handle rather than something else (e.g. a lock handle).

Also rename the confusingly-named mti_close_handle to mti_open_handle,
since this is referencing an open file handle, even if it is used at
close time to close the file.

   mfd_handle -> mfd_open_handle
   mfd_old_handle -> mfd_open_handle_old
   mfd_handle2mfd() -> mfd_open_handle2mfd()
   mti_close_handle -> mti_open_handle
   mbo_handle -> mbo_open_handle
   mio_handle -> mio_open_handle
   op_handle -> op_open_handle
   cr_old_handle -> cr_open_handle_old
   rr_handle -> rr_open_handle
   rs_handle -> rs_open_handle
   och_fh -> och_open_handle

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I6e2abdaee1c32d502ea4f2f4b339a6bb21528f69